+++ alidns-20150109/alibabacloud_alidns20150109/models.py
@@ -5340,6 +5340,493 @@ class DescribeBatchResultDetailResponse(TeaModel):
         return self
 
 
+class DescribeCloudGtmAddressPoolRequest(TeaModel):
+    def __init__(
+        self,
+        accept_language: str = None,
+        address_pool_id: str = None,
+        client_token: str = None,
+    ):
+        self.accept_language = accept_language
+        self.address_pool_id = address_pool_id
+        self.client_token = client_token
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.accept_language is not None:
+            result['AcceptLanguage'] = self.accept_language
+        if self.address_pool_id is not None:
+            result['AddressPoolId'] = self.address_pool_id
+        if self.client_token is not None:
+            result['ClientToken'] = self.client_token
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AcceptLanguage') is not None:
+            self.accept_language = m.get('AcceptLanguage')
+        if m.get('AddressPoolId') is not None:
+            self.address_pool_id = m.get('AddressPoolId')
+        if m.get('ClientToken') is not None:
+            self.client_token = m.get('ClientToken')
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasksHealthTask(TeaModel):
+    def __init__(
+        self,
+        port: int = None,
+        template_id: str = None,
+        template_name: str = None,
+    ):
+        self.port = port
+        self.template_id = template_id
+        self.template_name = template_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.port is not None:
+            result['Port'] = self.port
+        if self.template_id is not None:
+            result['TemplateId'] = self.template_id
+        if self.template_name is not None:
+            result['TemplateName'] = self.template_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Port') is not None:
+            self.port = m.get('Port')
+        if m.get('TemplateId') is not None:
+            self.template_id = m.get('TemplateId')
+        if m.get('TemplateName') is not None:
+            self.template_name = m.get('TemplateName')
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasks(TeaModel):
+    def __init__(
+        self,
+        health_task: List[DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasksHealthTask] = None,
+    ):
+        self.health_task = health_task
+
+    def validate(self):
+        if self.health_task:
+            for k in self.health_task:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['HealthTask'] = []
+        if self.health_task is not None:
+            for k in self.health_task:
+                result['HealthTask'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.health_task = []
+        if m.get('HealthTask') is not None:
+            for k in m.get('HealthTask'):
+                temp_model = DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasksHealthTask()
+                self.health_task.append(temp_model.from_map(k))
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBodyAddressesAddressRequestSource(TeaModel):
+    def __init__(
+        self,
+        request_source: List[str] = None,
+    ):
+        self.request_source = request_source
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.request_source is not None:
+            result['RequestSource'] = self.request_source
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('RequestSource') is not None:
+            self.request_source = m.get('RequestSource')
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBodyAddressesAddress(TeaModel):
+    def __init__(
+        self,
+        address: str = None,
+        address_id: str = None,
+        attribute_info: str = None,
+        available_mode: str = None,
+        available_status: str = None,
+        create_time: str = None,
+        create_timestamp: int = None,
+        enable_status: str = None,
+        health_judgement: str = None,
+        health_status: str = None,
+        health_tasks: DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasks = None,
+        manual_available_status: str = None,
+        name: str = None,
+        remark: str = None,
+        request_source: DescribeCloudGtmAddressPoolResponseBodyAddressesAddressRequestSource = None,
+        seq_non_preemptive_schedule: bool = None,
+        serial_number: int = None,
+        type: str = None,
+        update_time: str = None,
+        update_timestamp: int = None,
+        weight_value: int = None,
+    ):
+        self.address = address
+        self.address_id = address_id
+        self.attribute_info = attribute_info
+        self.available_mode = available_mode
+        self.available_status = available_status
+        self.create_time = create_time
+        self.create_timestamp = create_timestamp
+        self.enable_status = enable_status
+        self.health_judgement = health_judgement
+        self.health_status = health_status
+        self.health_tasks = health_tasks
+        self.manual_available_status = manual_available_status
+        self.name = name
+        self.remark = remark
+        self.request_source = request_source
+        self.seq_non_preemptive_schedule = seq_non_preemptive_schedule
+        self.serial_number = serial_number
+        self.type = type
+        self.update_time = update_time
+        self.update_timestamp = update_timestamp
+        self.weight_value = weight_value
+
+    def validate(self):
+        if self.health_tasks:
+            self.health_tasks.validate()
+        if self.request_source:
+            self.request_source.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.address is not None:
+            result['Address'] = self.address
+        if self.address_id is not None:
+            result['AddressId'] = self.address_id
+        if self.attribute_info is not None:
+            result['AttributeInfo'] = self.attribute_info
+        if self.available_mode is not None:
+            result['AvailableMode'] = self.available_mode
+        if self.available_status is not None:
+            result['AvailableStatus'] = self.available_status
+        if self.create_time is not None:
+            result['CreateTime'] = self.create_time
+        if self.create_timestamp is not None:
+            result['CreateTimestamp'] = self.create_timestamp
+        if self.enable_status is not None:
+            result['EnableStatus'] = self.enable_status
+        if self.health_judgement is not None:
+            result['HealthJudgement'] = self.health_judgement
+        if self.health_status is not None:
+            result['HealthStatus'] = self.health_status
+        if self.health_tasks is not None:
+            result['HealthTasks'] = self.health_tasks.to_map()
+        if self.manual_available_status is not None:
+            result['ManualAvailableStatus'] = self.manual_available_status
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.remark is not None:
+            result['Remark'] = self.remark
+        if self.request_source is not None:
+            result['RequestSource'] = self.request_source.to_map()
+        if self.seq_non_preemptive_schedule is not None:
+            result['SeqNonPreemptiveSchedule'] = self.seq_non_preemptive_schedule
+        if self.serial_number is not None:
+            result['SerialNumber'] = self.serial_number
+        if self.type is not None:
+            result['Type'] = self.type
+        if self.update_time is not None:
+            result['UpdateTime'] = self.update_time
+        if self.update_timestamp is not None:
+            result['UpdateTimestamp'] = self.update_timestamp
+        if self.weight_value is not None:
+            result['WeightValue'] = self.weight_value
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Address') is not None:
+            self.address = m.get('Address')
+        if m.get('AddressId') is not None:
+            self.address_id = m.get('AddressId')
+        if m.get('AttributeInfo') is not None:
+            self.attribute_info = m.get('AttributeInfo')
+        if m.get('AvailableMode') is not None:
+            self.available_mode = m.get('AvailableMode')
+        if m.get('AvailableStatus') is not None:
+            self.available_status = m.get('AvailableStatus')
+        if m.get('CreateTime') is not None:
+            self.create_time = m.get('CreateTime')
+        if m.get('CreateTimestamp') is not None:
+            self.create_timestamp = m.get('CreateTimestamp')
+        if m.get('EnableStatus') is not None:
+            self.enable_status = m.get('EnableStatus')
+        if m.get('HealthJudgement') is not None:
+            self.health_judgement = m.get('HealthJudgement')
+        if m.get('HealthStatus') is not None:
+            self.health_status = m.get('HealthStatus')
+        if m.get('HealthTasks') is not None:
+            temp_model = DescribeCloudGtmAddressPoolResponseBodyAddressesAddressHealthTasks()
+            self.health_tasks = temp_model.from_map(m['HealthTasks'])
+        if m.get('ManualAvailableStatus') is not None:
+            self.manual_available_status = m.get('ManualAvailableStatus')
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('Remark') is not None:
+            self.remark = m.get('Remark')
+        if m.get('RequestSource') is not None:
+            temp_model = DescribeCloudGtmAddressPoolResponseBodyAddressesAddressRequestSource()
+            self.request_source = temp_model.from_map(m['RequestSource'])
+        if m.get('SeqNonPreemptiveSchedule') is not None:
+            self.seq_non_preemptive_schedule = m.get('SeqNonPreemptiveSchedule')
+        if m.get('SerialNumber') is not None:
+            self.serial_number = m.get('SerialNumber')
+        if m.get('Type') is not None:
+            self.type = m.get('Type')
+        if m.get('UpdateTime') is not None:
+            self.update_time = m.get('UpdateTime')
+        if m.get('UpdateTimestamp') is not None:
+            self.update_timestamp = m.get('UpdateTimestamp')
+        if m.get('WeightValue') is not None:
+            self.weight_value = m.get('WeightValue')
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBodyAddresses(TeaModel):
+    def __init__(
+        self,
+        address: List[DescribeCloudGtmAddressPoolResponseBodyAddressesAddress] = None,
+    ):
+        self.address = address
+
+    def validate(self):
+        if self.address:
+            for k in self.address:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['Address'] = []
+        if self.address is not None:
+            for k in self.address:
+                result['Address'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.address = []
+        if m.get('Address') is not None:
+            for k in m.get('Address'):
+                temp_model = DescribeCloudGtmAddressPoolResponseBodyAddressesAddress()
+                self.address.append(temp_model.from_map(k))
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponseBody(TeaModel):
+    def __init__(
+        self,
+        address_lb_strategy: str = None,
+        address_pool_id: str = None,
+        address_pool_name: str = None,
+        address_pool_type: str = None,
+        addresses: DescribeCloudGtmAddressPoolResponseBodyAddresses = None,
+        available_status: str = None,
+        create_time: str = None,
+        create_timestamp: int = None,
+        enable_status: str = None,
+        health_judgement: str = None,
+        health_status: str = None,
+        remark: str = None,
+        request_id: str = None,
+        sequence_lb_strategy_mode: str = None,
+        update_time: str = None,
+        update_timestamp: int = None,
+    ):
+        self.address_lb_strategy = address_lb_strategy
+        self.address_pool_id = address_pool_id
+        self.address_pool_name = address_pool_name
+        self.address_pool_type = address_pool_type
+        self.addresses = addresses
+        self.available_status = available_status
+        self.create_time = create_time
+        self.create_timestamp = create_timestamp
+        self.enable_status = enable_status
+        self.health_judgement = health_judgement
+        self.health_status = health_status
+        self.remark = remark
+        self.request_id = request_id
+        self.sequence_lb_strategy_mode = sequence_lb_strategy_mode
+        self.update_time = update_time
+        self.update_timestamp = update_timestamp
+
+    def validate(self):
+        if self.addresses:
+            self.addresses.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.address_lb_strategy is not None:
+            result['AddressLbStrategy'] = self.address_lb_strategy
+        if self.address_pool_id is not None:
+            result['AddressPoolId'] = self.address_pool_id
+        if self.address_pool_name is not None:
+            result['AddressPoolName'] = self.address_pool_name
+        if self.address_pool_type is not None:
+            result['AddressPoolType'] = self.address_pool_type
+        if self.addresses is not None:
+            result['Addresses'] = self.addresses.to_map()
+        if self.available_status is not None:
+            result['AvailableStatus'] = self.available_status
+        if self.create_time is not None:
+            result['CreateTime'] = self.create_time
+        if self.create_timestamp is not None:
+            result['CreateTimestamp'] = self.create_timestamp
+        if self.enable_status is not None:
+            result['EnableStatus'] = self.enable_status
+        if self.health_judgement is not None:
+            result['HealthJudgement'] = self.health_judgement
+        if self.health_status is not None:
+            result['HealthStatus'] = self.health_status
+        if self.remark is not None:
+            result['Remark'] = self.remark
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.sequence_lb_strategy_mode is not None:
+            result['SequenceLbStrategyMode'] = self.sequence_lb_strategy_mode
+        if self.update_time is not None:
+            result['UpdateTime'] = self.update_time
+        if self.update_timestamp is not None:
+            result['UpdateTimestamp'] = self.update_timestamp
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AddressLbStrategy') is not None:
+            self.address_lb_strategy = m.get('AddressLbStrategy')
+        if m.get('AddressPoolId') is not None:
+            self.address_pool_id = m.get('AddressPoolId')
+        if m.get('AddressPoolName') is not None:
+            self.address_pool_name = m.get('AddressPoolName')
+        if m.get('AddressPoolType') is not None:
+            self.address_pool_type = m.get('AddressPoolType')
+        if m.get('Addresses') is not None:
+            temp_model = DescribeCloudGtmAddressPoolResponseBodyAddresses()
+            self.addresses = temp_model.from_map(m['Addresses'])
+        if m.get('AvailableStatus') is not None:
+            self.available_status = m.get('AvailableStatus')
+        if m.get('CreateTime') is not None:
+            self.create_time = m.get('CreateTime')
+        if m.get('CreateTimestamp') is not None:
+            self.create_timestamp = m.get('CreateTimestamp')
+        if m.get('EnableStatus') is not None:
+            self.enable_status = m.get('EnableStatus')
+        if m.get('HealthJudgement') is not None:
+            self.health_judgement = m.get('HealthJudgement')
+        if m.get('HealthStatus') is not None:
+            self.health_status = m.get('HealthStatus')
+        if m.get('Remark') is not None:
+            self.remark = m.get('Remark')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('SequenceLbStrategyMode') is not None:
+            self.sequence_lb_strategy_mode = m.get('SequenceLbStrategyMode')
+        if m.get('UpdateTime') is not None:
+            self.update_time = m.get('UpdateTime')
+        if m.get('UpdateTimestamp') is not None:
+            self.update_timestamp = m.get('UpdateTimestamp')
+        return self
+
+
+class DescribeCloudGtmAddressPoolResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: DescribeCloudGtmAddressPoolResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = DescribeCloudGtmAddressPoolResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class DescribeCustomLineRequest(TeaModel):
     def __init__(
         self,